Funçoes Lambda
Criação das policys e roles para as funções lambda.
Criaçao das funções lambdas e alteração do nome das funções incorretas.
Inclusão da documentação criada automaticamente pela AWS.
Atualização da documentação de operação (README).
