+++ Lambda_function/deployget.py
