+++ Lambda_function/teste-deploystore.json
@@ -1,0 +1,6 @@
+{
+    "nomeComponente":  "Componente 1",
+    "numeroVersao": "Versao 1",
+    "nomeResponsavel": "Marcelo Vilela",
+    "statusDeploy": "Em andamento"
+}
